Unit testability: split out definition of config page structs.
